Corrected if/else statements, removed querious from install

+++ inti_applications.sh
@@ -2,87 +2,115 @@
 if [ -f "/usr/local/Homebrew/bin/brew" ]; then
 		brew update
 
-
 			# install navicat
-		fi [ ! -f "/Applications/Naviat for MySQL.app" ]; then
+		if [ "/Applications/Naviat for MySQL.app" ]; then
+			echo "Already installed!"
+		else
 			brew cask install navicat-for-mysql
 		fi
 
 		# Sequel Pro
-		fi [ ! -f "/Applications/Sequel Pro.app" ]; then
+		if [ -f "/Applications/Sequel Pro.app" ]; then
+			echo "Already installed!"
+		else
 			brew cask install sequel-pro
 		fi
 
 		# install 1Password
-		if [ ! -f "/Applications/1Password.app" ]; then
+		if [ -f "/Applications/1Password.app" ]; then
+			echo "Already installed!"
+		else
 			brew cask install 1password
 		fi
 
 		# install Kaleidoscope
-		if [ ! -f "/Applications/Kaleidoscope.app" ]; then
+		if [ -f "/Applications/Kaleidoscope.app" ]; then
+			echo "Already installed!"
+		else
 			brew cask install kaleidoscope
 		fi
 
-		# install Querious
-		if [ ! -f "/Applications/Querious.app" ]; then
-			brew cask install querious
-		fi
+		# # install Querious
+		# if [ -f "/Applications/Querious.app" ]; then
+		# 	echo "Already installed!"
+		# else
+		# 	brew cask install querious
+		# fi
 
 		# install slack
-		if [ ! -f "/Applications/Slack.app" ]; then
+		if [ -f "/Applications/Slack.app" ]; then
+			echo "Already installed!"
+		else
 			brew cask install slack
 		fi
 
 		# install screenHero
-		if [ ! -f "/Applications/Screenhero.app" ]; then
+		if [ -f "/Applications/Screenhero.app" ]; then
+			echo "Already installed!"
+		else
 			brew cask install screenhero
 		fi
 
 		# install dropbox
-		if [ ! -f "/Applications/Dropbox.app" ]; then
+		if [ -f "/Applications/Dropbox.app" ]; then
+			echo "Already installed!"
+		else
 			brew cask install dropbox
 		fi
 
 		# install vlc
-		if [ ! -f "/Applications/VLC.app" ]; then
+		if [ -f "/Applications/VLC.app" ]; then
+			echo "Already installed!"
+		else
 			brew cask install vlc
 		fi
 
 		# install skype
-		fi [ ! -f "/Applications/Skype.app" ]; then
+		if [ -f "/Applications/Skype.app" ]; then
+			echo "Already installed!"
+		else
 			brew cask install skype
 		fi
 
 		# install spotify
-		fi [ ! -f "/Applications/Spotify.app" ]; then
+		if [ -f "/Applications/Spotify.app" ]; then
+			echo "Already installed!"
+		else
 			brew cask install spotify
 		fi
 
 		# install Textual
-		fi [ ! -f "/Applications/Textual.app" ]; then
+		if [ -f "/Applications/Textual.app" ]; then
+			echo "Already installed!"
+		else
 			brew cask install textual
 		fi
 
 		# install Fantastical
-		fi [ ! -f "/Applications/Fantastical.app" ]; then
+		if [ -f "/Applications/Fantastical.app" ]; then
+			echo "Already installed!"
+		else
 			brew cask install fantastical
 		fi
 
 		# Install VirtualBox
-		fi [ ! -f "/Applications/VirtualBox.app" ]; then
+		if [ -f "/Applications/VirtualBox.app" ]; then
+			echo "Already installed!"
+		else
 			brew cask install virtualbox
 		fi
 
 		# nstall vagrant
-		fi [ ! -f "/opt/vagrant/bin/vagrant" ]; then
-		brew cask install vagrant
+		if [ -f "/opt/vagrant/bin/vagrant" ]; then
+			echo "Already installed!"
+		else
+			brew cask install vagrant
 		fi
 
 		# Install Vagrant-Mananger
-		fi [ ! -f "/Applications/Vagrant Manager.app" ]; then
+		if [ -f "/Applications/Vagrant Manager.app" ]; then
+			echo "Already installed!"
+		else
 			brew cask install vagrant-manager
 		fi
-
-
-
 fi
